+++ src/main/java/br/com/mini_erp/estoque/ProdutoService.java
@@ -1,6 +1,8 @@
 package br.com.mini_erp.estoque;
 
+import br.com.mini_erp.shared.exception.ResourceNotFoundException; // Importe
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 
 @Service
@@ -16,17 +18,22 @@ public class ProdutoService {
         return repository.findAll();
     }
 
+    @Transactional
     public Produto salvar(Produto produto) {
         return repository.save(produto);
     }
 
     public Produto buscarPorId(Long id) {
         return repository.findById(id).orElseThrow(() ->
-                new RuntimeException("Produto não encontrado")
+                new ResourceNotFoundException("Produto não encontrado com ID: " + id) // <<-- ALTERADO AQUI
         );
     }
 
+    @Transactional
     public void deletar(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Produto não encontrado com ID: " + id);
+        }
         repository.deleteById(id);
     }
 }
